+++ flake.nix
@@ -97,7 +97,7 @@
             rustup
             rustPlatform.bindgenHook
             pkg-config
-            (python3.withPackages (ps: with ps; [ numpy ]))
+            (python3.withPackages (ps: with ps; [ numpy skyfield ]))
           ];
           inherit buildInputs;
           RUSTC_VERSION = "stable";