+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **Average plotting**. By default, rSTRF plots in "max-hold" mode: if multiple spectrogram bins
   render to the same pixel, the pixel is displayed with the highest power of all the bins. Now,
   there is an alternative "average" mode that averages the bins together instead.
+- `pass_png_historic.py` now skips spectrograms with no passes in them.
 
 # v0.3.2